refactor(accordion): move accordion group into subfolder
