+++ src/components/accordion/accordion-group/accordion-group.component.js
@@ -1,10 +1,10 @@
 import React, { useMemo, useCallback } from "react";
 import styledSystemPropTypes from "@styled-system/prop-types";
 
-import { filterStyledSystemMarginProps } from "../../style/utils";
-import Events from "../../utils/helpers/events";
-import Accordion from "./accordion.component.js";
-import { StyledAccordionGroup } from "./accordion.style.js";
+import { filterStyledSystemMarginProps } from "../../../style/utils";
+import Events from "../../../utils/helpers/events";
+import Accordion from "../accordion.component.js";
+import { StyledAccordionGroup } from "../accordion.style.js";
 
 const marginProptypes = filterStyledSystemMarginProps(
   styledSystemPropTypes.space
