+++ src/components/accordion/accordion-group/accordion-group.d.ts
@@ -1,8 +1,12 @@
 import * as React from "react";
 import { MarginProps } from "styled-system";
-import { AccordionProps } from "./accordion";
+import { AccordionProps } from "../accordion";
 
-type AccordionElement = React.ReactElement<AccordionProps> | boolean | null | undefined;
+type AccordionElement =
+  | React.ReactElement<AccordionProps>
+  | boolean
+  | null
+  | undefined;
 
 export interface AccordionGroupProps extends MarginProps {
   children?: AccordionElement | AccordionElement[];
