+++ src/components/accordion/accordion.stories.js
@@ -3,7 +3,7 @@ import { select, text, number } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 import OptionsHelper from "../../utils/helpers/options-helper";
 import Accordion from "./accordion.component";
-import AccordionGroup from "./accordion-group.component";
+import AccordionGroup from "./accordion-group/accordion-group.component";
 import Textbox from "../../__experimental__/components/textbox";
 import Box from "../box";
