+++ src/components/accordion/index.d.ts
@@ -1,2 +1,2 @@
 export { default as Accordion } from "./accordion";
-export { default as AccordionGroup } from "./accordion-group";
+export { default as AccordionGroup } from "./accordion-group/accordion-group";
