+++ src/components/accordion/index.js
@@ -1,2 +1,2 @@
 export { default as Accordion } from "./accordion.component";
-export { default as AccordionGroup } from "./accordion-group.component";
+export { default as AccordionGroup } from "./accordion-group/accordion-group.component";
